Cloud_function/requirements.txt
functions-framework==3.*
google-cloud-compute<3.0.0